requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.3
django-cors-headers==4.2.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
drf-spectacular==0.26.4
drf-yasg==1.21.7
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
Markdown==3.4.4
packaging==23.1
pluggy==1.2.0
PyJWT==1.7.1
pytest==7.4.0
python-decouple==3.8
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.0
requests==2.31.0
rpds-py==0.9.2
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==2.0.4
